server/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-gis==1.0
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
redis==5.0.7
celery==5.3.1
python-decouple==3.8
djoser==2.2.2
django-extensions==3.2.3
Pillow==10.4.0